analytics-service/requirements.txt
flask==2.3.3
snowflake-connector-python==3.7.0
gunicorn==21.2.0
requests==2.31.0 